fix: 와인 렌더링 버그 수정
